Changed UserId management system (more coherent now). Turned many slots into simple functions (they were not used as slots and we want to limit the number of slots anyways because we suspect some crashes are coming from inter-thread slot calls). Removed some useless code. Protected also read access to nodes in scene. drawing window's requestRedraw is now a simple UpdateGL, which should be good enough according to the documentation (simply queues a redraw, don't really do it).
git-svn-id: file:///home/simon/NFBMigrationGit/CodeSvn/Code/trunk@34 d53ed5f0-1798-4b2c-8c60-f91484f69ffc

+++ DrawLib/drwNetworkClient.h
@@ -20,7 +20,7 @@ public:
 	drwNetworkClient( QString & userName, QHostAddress & address, drwCommandDispatcher * dispatcher );
 	~drwNetworkClient();
 
-	double GetPercentRead();
+	int GetPercentRead();
 
 	ClientState GetState() { return m_state; }
